Remove dead code (custom platform)
